Refactor hook implementation by introducing BaseHook class and moving recoil constants to MacroConstants
